Use cl-generic package
Depend on `cl-generic' and use `cl-defgeneric' and `cl-defmethod' that
it defines, instead of `defgeneric' and `defmethod' from the `eieio'
package.

`cl-generic' is part of Emacs since 25.1.  It defines the functions
`cl-defgeneric' and `cl-defmethod', which replace the prefix-less
variants `defgeneric' and `defmethod'.

The latter two are now considered obsolete and to allow authors to
move away from them while maintaining support for older Emacsen, GNU
Elpa features a forward-port of the `cl-generic' package, similar to
what done in the past with `cl-lib'.