Define emacsql-compile as a generic function
Do so because the new `sqlite3' back-end cannot use `emacsql-format'.
The substitute `emascql-format-partial' (to be added in the next
commit) will return a cons-cell instead of a string, and we cannot
`concat' to that.

It would probably be better to turn `emacsql-format' into a generic
function instead, and to move the `(concat ... ";")' from
`emacsql-compile' to the default method of that.