sqlite3: Use sqlite3_prepare_v2 and sqlite3_bind*
Like for other back-ends we continue to expand identifiers because SQL
statements cannot contain placeholders for identifiers. For parameters
however we use placeholders and bind the values using `sqlite3_bind*'.

The benefit of that is that SQL no longer tokenize the values, which
we need because we could otherwise trip over user data containing
weird characters.

Currently this breaks on a $v that stands for an identifiers (not
parameter) vector.  The next commit will fix that.