sqlite3: Add $V as a new placeholder type
This change has to be made in code also used by other back-ends,
but for those $v is just a synonym for $v.

The reason the sqlite3 backend needs $V in addition to $v is that
it would otherwise not be possible to tell vectors of identifiers
and vectors of parameters apart.  Unlike other back-ends sqlite3
needs to be able to do so because it does not expand parameter
vectors, but does expand identifier vectors like other back-ends.