Rename buffer type name to avoid compiler confusion (#78)
The variable name is the same as the type, causing the wrong value to
be returned from the buffer_create() function. The buffer_create()
caller would see buffer size 0, causing trouble when growing the
buffer when reading data ("middleware out of memory").

This addresses issue #78 which shows this error message in a Termux environment.