Show no warnings when dependencies are missing
Until now we distributed each back-end as a separate package,
which made sure that all of its dependencies are available.

The maintainers of NonGNU Elpa want to distribute all back-ends
as a single package.  We cannot make that package depend on all
dependencies of all the optional back-ends, so instead we have
to suppress warnings about missing dependencies.

Because some of the missing functions from back-end libraries are
implemented in C, `declare-function' does not succeed in convincing
`check-declare-directory'.  So we have to stop relying on the latter
as well.  It's all a bit unfortunate.