Rename "process" slot to "handle" and make emacsql-process obsolete
This slot was named "process" when all connector classes used a
subprocess to communicate with their databases.  Using a process is
fragile, which is the reason why `emacsql-sqlite-builtin-connection'
and `emacsql-sqlite-module-connection' were created.

The slot was not renamed because it was only intended for internal use
anyway, and renaming it comes with the risk of breaking packages that
accessed it directly anyway.

But to avoid confusion, rename it to `handle' now.  Also document what
types its value may have and that it is only intended for internal
use.  Make the `emacsql-process' accessor function obsolete, and forgo
providing a renamed substitute, to avoid advertising direct access to
the slot.  For a limited time only, continue to allow direct access to
the internal slot under its obsolete name, but make noise about it.

Closes #112.