emacsql-show-sql: Simplify how font-lock is ensured
We are in a newly created temporary buffer, so we do not need to
ensure that the buffer is non narrowed or that there is no outdated
fontification.  All we have to do is to ensure that the buffer is
fully font-locked before we insert the `buffer-string' into another
buffer.