Read all forms in column string returned by SQLite
Unlike the low-level function of the removed legacy SQLite back-end
(which inserted a single string into a buffer) `sqlite-select' and
`sqlite3-exec' return a two-dimensional tree, in which cells have
already been parsed.

When a cell is still a string after that, then we have to read a sexp
from that string.  That is usually enough, but in special cases the
string may contain multiple sexp, for example, given:

  (emacsql db [:insert-into tbl :values (["a"] ["b"] ["c"])])
  (emacsql db [:select (funcall group-concat tbl)])

we have to parse this string "\"a\"\";\"\"b\"\";\"\"c\"".

The legacy back-end did not have to parse such a string explicitly,
because it got to read the complete value (in this case confined to
the first line) in a buffer containing this instead:

  ((\"a\"\";\"\"b\"\";\"\"c\"))
  success
  #

so it can essentially do

  (car (read-from-string "((\"a\"\";\"\"b\"\";\"\"c\"))"))

to parse everything in one go.  The new back-ends have helpfully
made that this more complicated (which makes sense, it's just a
bit inconvenient, when using our "store everything as a string"
approach.

An approach for the new back-ends, which needs much less code,
could be:

  (caar (read-from-string (format "(%s)" col)))

but that feels less safe.

This commit teaches the new back-ends to behave the same as the
legacy back-end used to.  In the future we might want to look into
automatically stripping the separators.

Closes #127.